feat: add CustomerDeliveryLineQuantityCard component
