+++ packages/apps/stock/src/screens/customerDeliveries/CustomerDeliveryLineDetailScreen.js
@@ -17,28 +17,23 @@
  */
 
 import React, {useEffect, useState} from 'react';
-import {View, StyleSheet} from 'react-native';
 import {
-  Badge,
   HeaderContainer,
   Screen,
   ScrollView,
-  Text,
-  useThemeColor,
   NotesCard,
 } from '@axelor/aos-mobile-ui';
 import {useDispatch, useSelector, useTranslator} from '@axelor/aos-mobile-core';
 import {
-  QuantityCard,
   ProductCardInfo,
   StockMoveHeader,
   CustomerDeliveryLineDetailFixedItems,
+  CustomerDeliveryLineQuantityCard,
 } from '../../components';
 import {fetchProductWithId} from '../../features/productSlice';
 import StockMove from '../../types/stock-move';
 
 const CustomerDeliveryLineDetailScreen = ({route, navigation}) => {
-  const Colors = useThemeColor();
   const customerDelivery = route.params.customerDelivery;
   const customerDeliveryLine = route.params.customerDeliveryLine;
   const trackingNumber =
@@ -63,10 +58,6 @@ const CustomerDeliveryLineDetailScreen = ({route, navigation}) => {
       ),
     );
   }, [dispatch, customerDeliveryLine, route.params.product]);
-
-  const handleQtyChange = value => {
-    setRealQty(value);
-  };
 
   const handleShowProduct = () => {
     navigation.navigate('ProductStockDetailsScreen', {
@@ -114,43 +105,12 @@ const CustomerDeliveryLineDetailScreen = ({route, navigation}) => {
           trackingNumber={trackingNumber?.trackingNumberSeq}
           locker={customerDeliveryLine?.locker}
         />
-        <QuantityCard
-          labelQty={I18n.t('Stock_PickedQty')}
-          defaultValue={parseFloat(realQty).toFixed(2)}
-          onValueChange={handleQtyChange}
-          editable={
-            customerDelivery.statusSelect !== StockMove.status.Realized
-          }>
-          <View style={styles.headerQuantityCard}>
-            <Text style={styles.text}>
-              {`${I18n.t('Stock_AskedQty')} : ${parseFloat(
-                customerDeliveryLine != null ? customerDeliveryLine.qty : 0,
-              ).toFixed(2)} ${
-                customerDeliveryLine != null
-                  ? customerDeliveryLine.unit.name
-                  : product?.unit?.name
-              }`}
-            </Text>
-            {customerDeliveryLine != null && (
-              <View>
-                {Number(customerDeliveryLine.qty) !==
-                  Number(customerDeliveryLine.realQty) && (
-                  <Badge
-                    title={I18n.t('Stock_Status_Incomplete')}
-                    color={Colors.cautionColor}
-                  />
-                )}
-                {Number(customerDeliveryLine.qty) ===
-                  Number(customerDeliveryLine.realQty) && (
-                  <Badge
-                    title={I18n.t('Stock_Status_Complete')}
-                    color={Colors.primaryColor}
-                  />
-                )}
-              </View>
-            )}
-          </View>
-        </QuantityCard>
+        <CustomerDeliveryLineQuantityCard
+          customerDelivery={customerDelivery}
+          customerDeliveryLine={customerDeliveryLine}
+          realQty={realQty}
+          setRealQty={setRealQty}
+        />
         <NotesCard
           title={I18n.t('Stock_NotesClient')}
           data={customerDelivery.pickingOrderComments}
@@ -163,34 +123,5 @@ const CustomerDeliveryLineDetailScreen = ({route, navigation}) => {
     </Screen>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    marginVertical: '2%',
-  },
-  validateBtn: {
-    width: '60%',
-    marginTop: 10,
-    borderRadius: 35,
-    marginHorizontal: '20%',
-  },
-  stateLine: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginHorizontal: 32,
-    marginVertical: '1%',
-  },
-  stockView: {
-    marginTop: '2%',
-  },
-  text_secondary: {
-    fontSize: 14,
-  },
-  headerQuantityCard: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-});
 
 export default CustomerDeliveryLineDetailScreen;
